Calling the API from Signal Store when a zip code is added
